Update tests to match updated interfaces
